Added Review entity, created bases for DTOs for currently needed entities, created basis for mappers, created service interfaces with methods according to the current specification.

+++ impulse-backend/src/main/java/com/baseddevs/impulsebackend/repository/OrderItemRepository.java
@@ -4,5 +4,5 @@ import com.baseddevs.impulsebackend.model.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface OrderItemRepository extends JpaRepository<OrderItem, Integer>, JpaSpecificationExecutor<OrderItem> {
+public interface OrderItemRepository extends JpaRepository<OrderItem, Long>, JpaSpecificationExecutor<OrderItem> {
 }
