+++ impulse-backend/src/main/java/com/baseddevs/impulsebackend/repository/PromoCodeRepository.java
@@ -4,5 +4,5 @@ import com.baseddevs.impulsebackend.model.PromoCode;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface PromoCodeRepository extends JpaRepository<PromoCode, Integer>, JpaSpecificationExecutor<PromoCode> {
+public interface PromoCodeRepository extends JpaRepository<PromoCode, Long>, JpaSpecificationExecutor<PromoCode> {
 }
